+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 FEATURES:
 * **New Data Source:** sumologic_partition
+* **New Data Source:** sumologic_partitions
 
 ENHANCEMENTS:
 * Add support for searchable time for dashboard field time_source